+++ x-pack/plugins/ml/public/application/_index.scss
@@ -1,11 +1,6 @@
 // ML has it's own variables for coloring
 @import 'variables';
 
-// Kibana management page ML section
-#kibanaManagementMLSection {
-  @import 'management/index';
-}
-
 // Protect the rest of Kibana from ML generic namespacing
 // SASSTODO: Prefix ml selectors instead
 .ml-app {
@@ -24,7 +19,6 @@
   // Components
   @import 'components/annotations/annotation_description_list/index'; // SASSTODO: This file overwrites EUI directly
   @import 'components/anomalies_table/index'; // SASSTODO: This file overwrites EUI directly
-  @import 'components/chart_tooltip/index';
   @import 'components/color_range_legend/index';
   @import 'components/controls/index';
   @import 'components/entity_cell/index';
