--- x-pack/plugins/ml/public/application/management/_index.scss
@@ -1,1 +1,0 @@
-@import 'jobs_list/index';
