+++ x-pack/plugins/ml/public/application/management/jobs_list/_index.scss
@@ -1,1 +1,4 @@
-@import 'components/index';
+// Kibana management page ML section
+#kibanaManagementMLSection {
+  @import 'components/index';
+}
