+++ x-pack/plugins/ml/public/application/management/jobs_list/index.ts
@@ -12,6 +12,7 @@ import { MlStartDependencies } from '../../../plugin';
 import { JobsListPage } from './components';
 import { getJobsListBreadcrumbs } from '../breadcrumbs';
 import { setDependencyCache, clearCache } from '../../util/dependency_cache';
+import './_index.scss';
 
 const renderApp = (element: HTMLElement, coreStart: CoreStart) => {
   ReactDOM.render(React.createElement(JobsListPage, { coreStart }), element);
