+++ x-pack/plugins/ml/public/embeddables/anomaly_swimlane/anomaly_swimlane_embeddable.tsx
@@ -9,29 +9,17 @@ import ReactDOM from 'react-dom';
 import { CoreStart } from 'kibana/public';
 import { i18n } from '@kbn/i18n';
 import { Subject } from 'rxjs';
-import {
-  Embeddable,
-  EmbeddableInput,
-  EmbeddableOutput,
-  IContainer,
-  IEmbeddable,
-} from '../../../../../../src/plugins/embeddable/public';
+import { Embeddable, IContainer } from '../../../../../../src/plugins/embeddable/public';
 import { EmbeddableSwimLaneContainer } from './embeddable_swim_lane_container';
-import { AnomalyDetectorService } from '../../application/services/anomaly_detector_service';
 import { JobId } from '../../../common/types/anomaly_detection_jobs';
-import { AnomalyTimelineService } from '../../application/services/anomaly_timeline_service';
-import {
-  Filter,
-  Query,
-  RefreshInterval,
-  TimeRange,
-} from '../../../../../../src/plugins/data/common';
-import { SwimlaneType } from '../../application/explorer/explorer_constants';
 import { MlDependencies } from '../../application/app';
-import { AppStateSelectedCells } from '../../application/explorer/explorer_utils';
-import { SWIM_LANE_SELECTION_TRIGGER } from '../../ui_actions/triggers';
-
-export const ANOMALY_SWIMLANE_EMBEDDABLE_TYPE = 'ml_anomaly_swimlane';
+import { SWIM_LANE_SELECTION_TRIGGER } from '../../ui_actions';
+import {
+  ANOMALY_SWIMLANE_EMBEDDABLE_TYPE,
+  AnomalySwimlaneEmbeddableInput,
+  AnomalySwimlaneEmbeddableOutput,
+  AnomalySwimlaneServices,
+} from '..';
 
 export const getDefaultPanelTitle = (jobIds: JobId[]) =>
   i18n.translate('xpack.ml.swimlaneEmbeddable.title', {
@@ -39,51 +27,7 @@ export const getDefaultPanelTitle = (jobIds: JobId[]) =>
     values: { jobIds: jobIds.join(', ') },
   });
 
-export interface AnomalySwimlaneEmbeddableCustomInput {
-  jobIds: JobId[];
-  swimlaneType: SwimlaneType;
-  viewBy?: string;
-  perPage?: number;
-
-  // Embeddable inputs which are not included in the default interface
-  filters: Filter[];
-  query: Query;
-  refreshConfig: RefreshInterval;
-  timeRange: TimeRange;
-}
-
-export interface EditSwimlanePanelContext {
-  embeddable: IEmbeddable<AnomalySwimlaneEmbeddableInput, AnomalySwimlaneEmbeddableOutput>;
-}
-
-export interface SwimLaneDrilldownContext extends EditSwimlanePanelContext {
-  /**
-   * Optional data provided by swim lane selection
-   */
-  data?: AppStateSelectedCells;
-}
-
-export type AnomalySwimlaneEmbeddableInput = EmbeddableInput & AnomalySwimlaneEmbeddableCustomInput;
-
-export type AnomalySwimlaneEmbeddableOutput = EmbeddableOutput &
-  AnomalySwimlaneEmbeddableCustomOutput;
-
-export interface AnomalySwimlaneEmbeddableCustomOutput {
-  perPage?: number;
-  fromPage?: number;
-  interval?: number;
-}
-
-export interface AnomalySwimlaneServices {
-  anomalyDetectorService: AnomalyDetectorService;
-  anomalyTimelineService: AnomalyTimelineService;
-}
-
-export type AnomalySwimlaneEmbeddableServices = [
-  CoreStart,
-  MlDependencies,
-  AnomalySwimlaneServices
-];
+export type IAnomalySwimlaneEmbeddable = typeof AnomalySwimlaneEmbeddable;
 
 export class AnomalySwimlaneEmbeddable extends Embeddable<
   AnomalySwimlaneEmbeddableInput,
