+++ x-pack/plugins/ml/public/embeddables/anomaly_swimlane/anomaly_swimlane_embeddable_factory.ts
@@ -10,23 +10,16 @@ import { StartServicesAccessor } from 'kibana/public';
 
 import {
   EmbeddableFactoryDefinition,
-  ErrorEmbeddable,
   IContainer,
 } from '../../../../../../src/plugins/embeddable/public';
-import {
-  ANOMALY_SWIMLANE_EMBEDDABLE_TYPE,
-  AnomalySwimlaneEmbeddable,
-  AnomalySwimlaneEmbeddableInput,
-  AnomalySwimlaneEmbeddableServices,
-} from './anomaly_swimlane_embeddable';
 import { HttpService } from '../../application/services/http_service';
-import { AnomalyDetectorService } from '../../application/services/anomaly_detector_service';
-import { AnomalyTimelineService } from '../../application/services/anomaly_timeline_service';
-import { mlResultsServiceProvider } from '../../application/services/results_service';
-import { resolveAnomalySwimlaneUserInput } from './anomaly_swimlane_setup_flyout';
-import { mlApiServicesProvider } from '../../application/services/ml_api_service';
 import { MlPluginStart, MlStartDependencies } from '../../plugin';
 import { MlDependencies } from '../../application/app';
+import {
+  ANOMALY_SWIMLANE_EMBEDDABLE_TYPE,
+  AnomalySwimlaneEmbeddableInput,
+  AnomalySwimlaneEmbeddableServices,
+} from '..';
 
 export class AnomalySwimlaneEmbeddableFactory
   implements EmbeddableFactoryDefinition<AnomalySwimlaneEmbeddableInput> {
@@ -50,6 +43,7 @@ export class AnomalySwimlaneEmbeddableFactory
     const [coreStart] = await this.getServices();
 
     try {
+      const { resolveAnomalySwimlaneUserInput } = await import('./anomaly_swimlane_setup_flyout');
       return await resolveAnomalySwimlaneUserInput(coreStart);
     } catch (e) {
       return Promise.reject();
@@ -58,6 +52,15 @@ export class AnomalySwimlaneEmbeddableFactory
 
   private async getServices(): Promise<AnomalySwimlaneEmbeddableServices> {
     const [coreStart, pluginsStart] = await this.getStartServices();
+
+    const { AnomalyDetectorService } = await import(
+      '../../application/services/anomaly_detector_service'
+    );
+    const { AnomalyTimelineService } = await import(
+      '../../application/services/anomaly_timeline_service'
+    );
+    const { mlApiServicesProvider } = await import('../../application/services/ml_api_service');
+    const { mlResultsServiceProvider } = await import('../../application/services/results_service');
 
     const httpService = new HttpService(coreStart.http);
     const anomalyDetectorService = new AnomalyDetectorService(httpService);
@@ -77,8 +80,9 @@ export class AnomalySwimlaneEmbeddableFactory
   public async create(
     initialInput: AnomalySwimlaneEmbeddableInput,
     parent?: IContainer
-  ): Promise<AnomalySwimlaneEmbeddable | ErrorEmbeddable> {
+  ): Promise<any> {
     const services = await this.getServices();
+    const { AnomalySwimlaneEmbeddable } = await import('./anomaly_swimlane_embeddable');
     return new AnomalySwimlaneEmbeddable(initialInput, services, parent);
   }
 }
