+++ x-pack/plugins/ml/public/embeddables/index.ts
@@ -8,6 +8,9 @@ import { AnomalySwimlaneEmbeddableFactory } from './anomaly_swimlane';
 import { MlCoreSetup } from '../plugin';
 import { EmbeddableSetup } from '../../../../../src/plugins/embeddable/public';
 
+export * from './constants';
+export * from './types';
+
 export function registerEmbeddables(embeddable: EmbeddableSetup, core: MlCoreSetup) {
   const anomalySwimlaneEmbeddableFactory = new AnomalySwimlaneEmbeddableFactory(
     core.getStartServices
