--- x-pack/plugins/ml/public/index.scss
@@ -1,1 +1,0 @@
-@import './application/index';
