+++ x-pack/plugins/ml/public/plugin.ts
@@ -6,36 +6,35 @@
 
 import { i18n } from '@kbn/i18n';
 import {
-  Plugin,
-  CoreStart,
-  CoreSetup,
   AppMountParameters,
+  CoreSetup,
+  CoreStart,
+  Plugin,
   PluginInitializerContext,
 } from 'kibana/public';
 import { BehaviorSubject } from 'rxjs';
 import { take } from 'rxjs/operators';
 import { ManagementSetup } from 'src/plugins/management/public';
-import { SharePluginSetup, SharePluginStart, UrlGeneratorState } from 'src/plugins/share/public';
+import { SharePluginSetup, SharePluginStart } from 'src/plugins/share/public';
 import { UsageCollectionSetup } from 'src/plugins/usage_collection/server';
 
 import { DataPublicPluginStart } from 'src/plugins/data/public';
 import { HomePublicPluginSetup } from 'src/plugins/home/public';
 import { EmbeddableSetup } from 'src/plugins/embeddable/public';
-import { AppStatus, AppUpdater } from '../../../../src/core/public';
+import { AppStatus, AppUpdater, DEFAULT_APP_CATEGORIES } from '../../../../src/core/public';
 import { SecurityPluginSetup } from '../../security/public';
 import { LicensingPluginSetup } from '../../licensing/public';
 import { registerManagementSection } from './application/management';
 import { LicenseManagementUIPluginSetup } from '../../license_management/public';
 import { setDependencyCache } from './application/util/dependency_cache';
-import { PLUGIN_ID, PLUGIN_ICON } from '../common/constants/app';
+import { PLUGIN_ICON, PLUGIN_ID } from '../common/constants/app';
 import { registerFeature } from './register_feature';
-import { DEFAULT_APP_CATEGORIES } from '../../../../src/core/public';
-import { registerEmbeddables } from './embeddables';
 import { UiActionsSetup, UiActionsStart } from '../../../../src/plugins/ui_actions/public';
 import { registerMlUiActions } from './ui_actions';
 import { KibanaLegacyStart } from '../../../../src/plugins/kibana_legacy/public';
-import { registerUrlGenerator, MlUrlGeneratorState, ML_APP_URL_GENERATOR } from './url_generator';
-import { isMlEnabled, isFullLicense } from '../common/license';
+import { registerUrlGenerator } from './url_generator';
+import { isFullLicense, isMlEnabled } from '../common/license';
+import { registerEmbeddables } from './embeddables';
 
 export interface MlStartDependencies {
   data: DataPublicPluginStart;
@@ -54,12 +53,6 @@ export interface MlSetupDependencies {
   uiActions: UiActionsSetup;
   kibanaVersion: string;
   share: SharePluginSetup;
-}
-
-declare module '../../../../src/plugins/share/public' {
-  export interface UrlGeneratorStateMapping {
-    [ML_APP_URL_GENERATOR]: UrlGeneratorState<MlUrlGeneratorState>;
-  }
 }
 
 export type MlCoreSetup = CoreSetup<MlStartDependencies, MlPluginStart>;
