+++ x-pack/plugins/ml/public/ui_actions/apply_influencer_filters_action.tsx
@@ -6,13 +6,10 @@
 
 import { i18n } from '@kbn/i18n';
 import { ActionContextMapping, createAction } from '../../../../../src/plugins/ui_actions/public';
-import {
-  AnomalySwimlaneEmbeddable,
-  SwimLaneDrilldownContext,
-} from '../embeddables/anomaly_swimlane/anomaly_swimlane_embeddable';
 import { MlCoreSetup } from '../plugin';
 import { SWIMLANE_TYPE, VIEW_BY_JOB_LABEL } from '../application/explorer/explorer_constants';
 import { Filter, FilterStateStore } from '../../../../../src/plugins/data/common';
+import { ANOMALY_SWIMLANE_EMBEDDABLE_TYPE, SwimLaneDrilldownContext } from '../embeddables';
 
 export const APPLY_INFLUENCER_FILTERS_ACTION = 'applyInfluencerFiltersAction';
 
@@ -73,7 +70,7 @@ export function createApplyInfluencerFiltersAction(
     async isCompatible({ embeddable, data }: SwimLaneDrilldownContext) {
       // Only compatible with view by influencer swim lanes and single selection
       return (
-        embeddable instanceof AnomalySwimlaneEmbeddable &&
+        embeddable.type === ANOMALY_SWIMLANE_EMBEDDABLE_TYPE &&
         data !== undefined &&
         data.type === SWIMLANE_TYPE.VIEW_BY &&
         data.viewByFieldName !== VIEW_BY_JOB_LABEL &&
