+++ x-pack/plugins/ml/public/ui_actions/index.ts
@@ -13,7 +13,6 @@ import {
   createOpenInExplorerAction,
   OPEN_IN_ANOMALY_EXPLORER_ACTION,
 } from './open_in_anomaly_explorer_action';
-import { EditSwimlanePanelContext } from '../embeddables/anomaly_swimlane/anomaly_swimlane_embeddable';
 import { UiActionsSetup } from '../../../../../src/plugins/ui_actions/public';
 import { MlPluginStart, MlStartDependencies } from '../plugin';
 import { CONTEXT_MENU_TRIGGER } from '../../../../../src/plugins/embeddable/public';
@@ -22,11 +21,18 @@ import {
   createApplyInfluencerFiltersAction,
 } from './apply_influencer_filters_action';
 import { SWIM_LANE_SELECTION_TRIGGER, swimLaneSelectionTrigger } from './triggers';
-import { SwimLaneDrilldownContext } from '../embeddables/anomaly_swimlane/anomaly_swimlane_embeddable';
 import {
   APPLY_TIME_RANGE_SELECTION_ACTION,
   createApplyTimeRangeSelectionAction,
 } from './apply_time_range_action';
+import { EditSwimlanePanelContext, SwimLaneDrilldownContext } from '../embeddables';
+
+export { APPLY_TIME_RANGE_SELECTION_ACTION } from './apply_time_range_action';
+export { EDIT_SWIMLANE_PANEL_ACTION } from './edit_swimlane_panel_action';
+export { APPLY_INFLUENCER_FILTERS_ACTION } from './apply_influencer_filters_action';
+export { OPEN_IN_ANOMALY_EXPLORER_ACTION } from './open_in_anomaly_explorer_action';
+
+export { SWIM_LANE_SELECTION_TRIGGER } from './triggers';
 
 /**
  * Register ML UI actions
