+++ x-pack/plugins/ml/public/ui_actions/open_in_anomaly_explorer_action.tsx
@@ -6,12 +6,9 @@
 
 import { i18n } from '@kbn/i18n';
 import { ActionContextMapping, createAction } from '../../../../../src/plugins/ui_actions/public';
-import {
-  AnomalySwimlaneEmbeddable,
-  SwimLaneDrilldownContext,
-} from '../embeddables/anomaly_swimlane/anomaly_swimlane_embeddable';
 import { MlCoreSetup } from '../plugin';
 import { ML_APP_URL_GENERATOR } from '../url_generator';
+import { ANOMALY_SWIMLANE_EMBEDDABLE_TYPE, SwimLaneDrilldownContext } from '../embeddables';
 
 export const OPEN_IN_ANOMALY_EXPLORER_ACTION = 'openInAnomalyExplorerAction';
 
@@ -60,7 +57,7 @@ export function createOpenInExplorerAction(getStartServices: MlCoreSetup['getSta
       await application.navigateToUrl(anomalyExplorerUrl!);
     },
     async isCompatible({ embeddable }: SwimLaneDrilldownContext) {
-      return embeddable instanceof AnomalySwimlaneEmbeddable;
+      return embeddable.type === ANOMALY_SWIMLANE_EMBEDDABLE_TYPE;
     },
   });
 }
